refactor: remove consumer
