--- shared/pkg/mq/constant.go
@@ -1,13 +1,0 @@
-package mq
-
-const (
-	Namespace = "rss3-pre-god-ns"
-	GroupName = "rss3-pre-god-group"
-)
-
-type TopicName string
-
-const (
-	TopicNameUpdateAssetDetail TopicName = "asset-detail"
-	TopicNameGiveAssetDetail   TopicName = "give-asset-detail"
-)
